Add subscriber to tests

+++ src/main/java/edu/kis/powp/jobs2d/drivers/UsageSubscriber.java
@@ -22,7 +22,7 @@ public class UsageSubscriber implements Subscriber {
         this.allOperations = allOperations;
     }
 
-    public void setDistance(float distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
